+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # SJHLocationManager CHANGELOG
 
+## 0.1.0
+
+Added support to use a delegate to receive location updates instead of notifications.
+
+## 0.0.4
+
+Complete CLLocationManagerDelegate implementation.
+
+## 0.0.3
 ## 0.0.2
-
-
 
 ## 0.0.1